Refactor determineProminentWords helper file
