+++ packages/yoastseo/spec/stringProcessing/determineProminentWordsEnglishSpec.js
@@ -1,5 +1,5 @@
 import ProminentWord from "../../src/values/ProminentWord";
-import { getProminentWords } from "../../src/languageProcessing/helpers/_todo/determineProminentWords";
+import { getProminentWords } from "../../src/languageProcessing/helpers/prominentWords/determineProminentWords";
 import getMorphologyData from "../specHelpers/getMorphologyData";
