+++ packages/yoastseo/spec/stringProcessing/determineProminentWordsPortugueseSpec.js
@@ -1,5 +1,5 @@
 import ProminentWord from "../../src/values/ProminentWord";
-import { getProminentWords } from "../../src/languageProcessing/helpers/_todo/determineProminentWords";
+import { getProminentWords } from "../../src/languageProcessing/helpers/prominentWords/determineProminentWords";
 
 describe( "gets Portuguese prominent words", function() {
 	it( "returns prominent words", function() {
